Removed namespace.yaml and use release namespace
Use the release namespace value for assigning the namespace in the templates. The user will be required to specify the namespace on the cli when doing a helm install using helm install --create-namespace -n tenable. The good thing about this is that all helm meta data gets stored in that namespace also rather than in the default namespace.